+++ test/fixtures/project_unused_decl/usingnamespace_main.zig
@@ -1,0 +1,5 @@
+usingnamespace @import("usingnamespace_api.zig");
+
+pub fn main() void {
+    exposed();
+}